pyproject.toml
[tool.poetry]
name = "mdurocherart"
version = "0.0.1"
description = ""
authors = ["Kasonde Kasonde <d.kasonde99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3"
pydantic = {version = "^2.2", extras =["email"]}
flask-mailman = "^0.3.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
bcrypt = "^4.0.1"


[tool.poetry.group.test.dependencies]
pytest = "7.4.0"